requirements.txt
mistralai
python-dotenv